Fixed Bengali Probhat layout, replaced duplicate keys with missing keys
Changed duplicate keys ঈ and ঠ for Probhat layout. The current version of Probhat keybpard of Indic-Keyboard contains "ধ ঊ ঈ ঈ ঠ ঠ উ ই ঔ ফ", which should be "ধ ঊ ঈ ড় ঠ ঐ উ ই ঔ ফ". Key for ড় and ঐ was missing. I replaced the duplicate keys with the missing keys.

Take a look at the official site of Probhat layout. http://ekushey.org/?page/probhat_layout
And here is an image of current buggy layout. http://i.imgur.com/CFotSBQ.png