Fix user submitted bugs: Bengali Dari is missing #54 DEVANAGARI DANDA instead of full stop in Hindi #26
Add Double Daari (॥) symbol to popup suggestions for Daari.
The position after # was chosen because its close enough to access for Bengali users while not getting in the way of hindi users.

(Force pushed to remove previous ~80 lines of code and new layout files that were created in attempt to fix this bug.
Now the fix is simplified to only a few additions. )